+++ sites/public/pages/listings.tsx
@@ -1,121 +1,15 @@
 import Head from "next/head"
 import qs from "qs"
 import axios from "axios"
-import {
-  ListingsGroup,
-  PageHeader,
-  t,
-  StatusBarType,
-  ApplicationStatusType,
-  ListingCard,
-  imageUrlFromListing,
-  getSummariesTable,
-} from "@bloom-housing/ui-components"
-import {
-  Listing,
-  ListingStatus,
-  ListingReviewOrder,
-  Address,
-  UnitsSummarized,
-} from "@bloom-housing/backend-core/types"
+import { ListingsGroup, PageHeader, t } from "@bloom-housing/ui-components"
+import { Listing, ListingStatus } from "@bloom-housing/backend-core/types"
 import Layout from "../layouts/application"
 import { MetaTags } from "../src/MetaTags"
-import moment from "moment"
-import { openInFuture } from "../lib/helpers"
+import { getListings } from "../lib/helpers"
 
 export interface ListingsProps {
   openListings: Listing[]
   closedListings: Listing[]
-}
-
-const getListingImageCardStatus = (listing: Listing): StatusBarType => {
-  let content = ""
-  let subContent = ""
-  let formattedDate = ""
-  let appStatus = ApplicationStatusType.Open
-
-  if (openInFuture(listing)) {
-    const date = listing.applicationOpenDate
-    const openDate = moment(date)
-    formattedDate = openDate.format("MMM. D, YYYY")
-    content = t("listings.applicationOpenPeriod")
-  } else {
-    if (listing.applicationDueDate) {
-      const dueDate = moment(listing.applicationDueDate)
-      const dueTime = moment(listing.applicationDueTime)
-      formattedDate = dueDate.format("MMM. DD, YYYY")
-
-      if (listing.applicationDueTime) {
-        formattedDate = formattedDate + ` ${t("t.at")} ` + dueTime.format("h:mm A")
-      }
-
-      // if due date is in future, listing is open
-      if (moment() < dueDate) {
-        content = t("listings.applicationDeadline")
-      } else {
-        appStatus = ApplicationStatusType.Closed
-        content = t("listings.applicationsClosed")
-      }
-    }
-  }
-
-  if (formattedDate != "") {
-    content = content + `: ${formattedDate}`
-  }
-
-  if (listing.reviewOrderType === ListingReviewOrder.firstComeFirstServe) {
-    subContent = content
-    content = t("listings.applicationFCFS")
-  }
-
-  return {
-    status: appStatus,
-    content,
-    subContent,
-  }
-}
-
-const getListingCardSubtitle = (address: Address) => {
-  const { street, city, state, zipCode } = address || {}
-  return address ? `${street}, ${city} ${state}, ${zipCode}` : null
-}
-
-const getListingTableData = (unitsSummarized: UnitsSummarized) => {
-  return unitsSummarized !== undefined ? getSummariesTable(unitsSummarized.byUnitTypeAndRent) : []
-}
-
-const getListings = (listings) => {
-  const unitSummariesHeaders = {
-    unitType: t("t.unitType"),
-    minimumIncome: t("t.minimumIncome"),
-    rent: t("t.rent"),
-  }
-  return listings.map((listing: Listing, index) => {
-    return (
-      <ListingCard
-        key={index}
-        imageCardProps={{
-          imageUrl:
-            imageUrlFromListing(listing, parseInt(process.env.listingPhotoSize || "1302")) || "",
-          subtitle: getListingCardSubtitle(listing.buildingAddress),
-          title: listing.name,
-          href: `/listing/${listing.id}/${listing.urlSlug}`,
-          tagLabel: listing.reservedCommunityType
-            ? t(`listings.reservedCommunityTypes.${listing.reservedCommunityType.name}`)
-            : undefined,
-          statuses: [getListingImageCardStatus(listing)],
-        }}
-        tableProps={{
-          headers: unitSummariesHeaders,
-          data: getListingTableData(listing.unitsSummarized),
-          responsiveCollapse: true,
-          cellClassName: "px-5 py-3",
-        }}
-        seeDetailsLink={`/listing/${listing.id}/${listing.urlSlug}`}
-        tableHeader={listing.showWaitlist ? t("listings.waitlist.open") : null}
-      />
-    )
-  })
 }
 
 const openListings = (listings) => {
